+++ app/models/character/companion.rb
@@ -2,6 +2,8 @@
 
 class Character
   class Companion < ApplicationRecord
+    has_one_attached :avatar
+
     belongs_to :character
   end
 end