+++ spec/factories/feats.rb
@@ -91,6 +91,25 @@ FactoryBot.define do
       kind { 0 }
     end
 
+    trait :pathfinder2 do
+      type { 'Pathfinder2::Feat' }
+      sequence(:slug) { |i| "slug-#{i}" }
+      title {
+        {
+          en: 'Name',
+          ru: 'Название'
+        }
+      }
+      description {
+        {
+          en: 'Description',
+          ru: 'Описание'
+        }
+      }
+      origin { 2 }
+      kind { 0 }
+    end
+
     trait :dnd2024_bardic_inspiration do
       type { 'Dnd2024::Feat' }
       slug { 'bardic_inspiration' }